requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.0
opencv-python==4.10.0.84
pyserial==3.5
Werkzeug==3.0.4
